% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra_subset.R
\name{add_coord}
\alias{add_coord}
\title{Add longitude and latitude of 'locid'}
\usage{
add_coord(x, force = FALSE)
}
\arguments{
\item{x}{}

\item{force}{}
}
\description{
Add longitude and latitude of 'locid'
}
\examples{
NA
}
