% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merra_subset.R
\name{add_lonlat}
\alias{add_lonlat}
\title{Add longitude and latitude of 'locid'}
\usage{
add_lonlat(x, force = FALSE)
}
\arguments{
\item{x}{}

\item{force}{}
}
\value{

}
\description{
Add longitude and latitude of 'locid'
}
