% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra_subset.R
\name{add_merra2_grid}
\alias{add_merra2_grid}
\title{Adds MERRA2 grid (polygons) to the data}
\usage{
add_merra2_grid(x)
}
\arguments{
\item{x}{data.frame with `locid` column}
}
\value{
`x` as `sf` object, with added geometry (polygons)
}
\description{
Adds MERRA2 grid (polygons) to the data
}
