% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra_subset.R
\name{add_merra2_points}
\alias{add_merra2_points}
\title{Adds MERRA2 grid (points) to the data}
\usage{
add_merra2_points(x)
}
\arguments{
\item{x}{data.frame with `locid` column}
}
\value{
`x` as `sf` object, with added geometry (points)
}
\description{
Adds MERRA2 grid (points) to the data
}
