% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{angle_of_incidence}
\alias{angle_of_incidence}
\alias{fAOI}
\title{Angle of Incidence (AOI)}
\usage{
angle_of_incidence(
  x,
  type = "fx",
  add.suffix = TRUE,
  azimuth = NULL,
  zenith = NULL,
  array.tilt = NULL,
  array.azimuth = NULL,
  na.val = NA,
  zenith.max = 90,
  AOI.max = 90,
  rad = TRUE
)

fAOI(
  x,
  type = "fx",
  add.suffix = TRUE,
  azimuth = NULL,
  zenith = NULL,
  array.tilt = NULL,
  array.azimuth = NULL,
  na.val = NA,
  zenith.max = 90,
  AOI.max = 90,
  rad = TRUE
)
}
\arguments{
\item{azimuth}{the solar zenith angle, degrees}

\item{zenith}{the solar azimuth angle, degrees}

\item{array.tilt}{the PV tilt angle, degrees}

\item{array.azimuth}{the PV azimuth angle, degrees}
}
\value{

}
\description{
Angle of Incidence (AOI)
}
\details{
\loadmathjax
    \mjsdeqn{AOI = \arccos\big({\cos{(zenith)}\cos{(array.tilt)}+
             \sin{(zenith)}\sin{(array.tilt)}\cos{(azimuth-array.azimuth)}\big)}}
    Though the equation returns AOI values with any given set of zenith, azimuth, 
    array.tilt, and array.azimuth, the AOI is meaningful only if 
    \mjseqn{0 \leq zenith \leq 90}, and \mjseqn{0 \leq AOI \leq 90}. 
    The former condition assures the sun is above the horizon and 
    the latter condition assures the sunlight beam is able to hit the panel
    (Source: Stackhouse et al., 2018).
}
\examples{
NA
}
