% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azi.conversion.R
\name{azimuth_N2Q}
\alias{azimuth_N2Q}
\title{Title}
\usage{
azimuth_N2Q(
  x,
  azimuth_N = "azimuth_N",
  azimuth_Q = "azimuth_Q",
  lat = "lat",
  lon = "lon"
)
}
\arguments{
\item{x}{}

\item{azimuth_N}{}

\item{azimuth_Q}{}

\item{lat}{}

\item{lon}{}
}
\description{
Title
}
