% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locid.R
\name{closest_locid}
\alias{closest_locid}
\title{Get MERRA-2 grid IDs closest to the given coordinates}
\usage{
closest_locid(lon, lat, asList = FALSE)
}
\arguments{
\item{lon}{longitude in degrees (-180 <= lon <= 180)}

\item{lat}{latitude in degrees (-90 <= lat <= 90)}

\item{asList}{}
}
\value{
integer vector with locations IDs (locid) when `asList` is FALSE (default). In the case of several values, only the first `locid` will be returned. If `asList` is TRUE, a list is returned with possible multiple values for each coordinate.
}
\description{
Get MERRA-2 grid IDs closest to the given coordinates
}
\examples{
closest_locid(0, 0)
closest_locid(100.14, -85.145)
closest_locid(0, 89.5, asList = TRUE)

}
