% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrel_pvwatts_api.R
\name{coord2tz_offset}
\alias{coord2tz_offset}
\title{Title}
\usage{
coord2tz_offset(lon, lat, utc = NULL)
}
\arguments{
\item{lon}{}

\item{lat}{}

\item{utc}{}
}
\value{

}
\description{
Title
}
\examples{
NA
}
