% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{deg2rad}
\alias{deg2rad}
\alias{rad2deg}
\title{Convert degrees to radians}
\usage{
deg2rad(x)

rad2deg(x)
}
\arguments{
\item{x}{}
}
\value{
numeric vector, radians
}
\description{
Convert degrees to radians
}
\examples{
rad2deg(pi)
deg2rad(180)
deg2rad(rad2deg(pi))
cos(pi); cos(deg2rad(rad2deg(pi)))
}
