% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrel_pvwatts_api.R
\name{fetch_pvwatts}
\alias{fetch_pvwatts}
\title{Title}
\usage{
fetch_pvwatts(
  query = list(lat = 0, lon = 0, dataset = "intl", radius = 1000, system_capacity = 1,
    module_type = 1, losses = 10, array_type = 0, tilt = 0, azimuth = 0, timeframe =
    "hourly"),
  array.type = NULL,
  ...,
  asList = TRUE,
  asJSON = !asList,
  pvwatts_api_url = "https://developer.nrel.gov/api/pvwatts/v6.json?api_key=",
  api_key = getOption("pwatts.api.key")
)
}
\arguments{
\item{query}{}

\item{array.type}{}

\item{...}{}

\item{asList}{}

\item{asJSON}{}

\item{pvwatts_api_url}{}

\item{api_key}{}
}
\value{

}
\description{
Title
}
\details{
<https://developer.nrel.gov/docs/solar/pvwatts/v6/>
}
\examples{
NA
}
