% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_power.R
\name{fHellmann}
\alias{fHellmann}
\alias{fH}
\title{Hellmann function to extrapolate wind speed from 10 and 50 meters to higher altitudes}
\usage{
fHellmann(W10M, W50M, na = 0, inf = 0)

fH(W10M, W50M, na = 0, inf = 0)
}
\arguments{
\item{W10M}{wind speed (m/s) at 10 meters altitude}

\item{W50M}{wind speed (m/s) at 50 meters altitude}

\item{na}{values to return in the case of NA}

\item{inf}{values to return instead of infinite}
}
\value{
a numeric vector with estimated Hellmann exponent.
}
\description{
Hellmann function to extrapolate wind speed from 10 and 50 meters to higher altitudes
}
\examples{
fH(5, 10)
}
\seealso{
\code{\link{fWindSpeedExtrapolation}} (\code{\link{fWSE}}) for wind speed extrapolation
}
