% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_power.R
\name{fWindCF}
\alias{fWindCF}
\title{Title}
\usage{
fWindCF(
  x,
  height = 50,
  mps = paste0("W", height, "M"),
  return_name = paste0("win", height, "af"),
  hellmann = "hellmann",
  W10M = "W10M",
  W50M = "W50M",
  WPC = fWindPowerCurve,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{data frame with MERRA-2 subset}

\item{height}{height over ground}

\item{mps}{name of wind speed variable available in `x` or will be extrapolated using `fHellmann` and `fWSE` functions}

\item{return_name}{name of the variable, which will be added (or overwritten) to `x`}

\item{hellmann}{name of the variable with Hellmann constant, either available in `x` or will be calculated using `fHellmann` function}

\item{W10M}{name of the variable with wind speed at 10 meters height}

\item{W50M}{name of the variable with wind speed at 10 meters height}

\item{WPC}{name of the wind power capacity function}

\item{verbose}{if TRUE, the process will be reported}

\item{...}{additional parameters for `fHellmann`, `fWSE`, `WPC` functions}
}
\value{
`x` with added (or overwritten) column of wind power capacity factors; the name of the column is given by `return_name` parameter.
}
\description{
Title
}
\examples{
NA
}
