% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_power.R
\name{fWindSpeedExtrapolation}
\alias{fWindSpeedExtrapolation}
\alias{fWSE}
\title{Hellmann-constant-based extrapolation of wind speed for higher altitudes}
\usage{
fWindSpeedExtrapolation(m, W10M, hellmann)

fWSE(m, W10M, hellmann)
}
\arguments{
\item{m}{height (in meters) for which wind speed will be estimated;}

\item{W10M}{wind speed at height of 10 meters (from MERRA2 database);}

\item{hellmann}{Hellmann exponent (calculated with fHellmann function)}
}
\value{
a numeric vector of estimated wind speed (m/s) at the \code{m} height (meters).
}
\description{
Hellmann-constant-based extrapolation of wind speed for higher altitudes
}
\examples{
h <- fH(5, 7)
fWSE(50, 5, h)
fWSE(100, 5, h)
fWSE(seq(50, 200, 50), 5, h)
}
\references{
<https://en.wikipedia.org/wiki/Wind_gradient>
}
\seealso{
\code{\link{fHellmann}} (\code{\link{fH}}) for estimation of Hellmann exponent
}
