% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locid.R
\name{get_locid}
\alias{get_locid}
\title{Get MERRA-2 grid IDs which overlaps with the given spatial object}
\usage{
get_locid(
  x,
  method = "polygons",
  add_poles_points = F,
  return_grid = FALSE,
  ...
)
}
\arguments{
\item{x}{spatial (`sf`) object. Expected CRS is `EPSG:4326`, if different, it will be attempted to transform the grid to CRS of `x`, though intersection algorithms may not work well for some projections. Use `sf::st_transform(x, "EPSG:4326")` and `sf::st_make_valid(x)` to transform and correct geometries.}

\item{method}{character, "polygons" (default) or "points" grid to use in `sf::st_intersects` method.}

\item{...}{ignored}
}
\description{
Get MERRA-2 grid IDs which overlaps with the given spatial object
}
