% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra_subset.R
\name{get_merra2_subset}
\alias{get_merra2_subset}
\title{Get MERRA-2 subset}
\usage{
get_merra2_subset(
  locid = 1:207936L,
  from = "1980-01-01 00",
  to = "2020-12-31 23",
  tz = "UTC",
  cols = NULL,
  quiet = FALSE,
  rows_lim = 2 * 10^9,
  as_integers = FALSE
)
}
\arguments{
\item{locid}{integer vector with location \code{locid} (values from 1 to 207936)}

\item{from}{starting date and time of the subset in "YYYY-MM-DD HH" format}

\item{to}{ending date and time of the subset in "YYYY-MM-DD HH" format (see \code{fDate})}

\item{tz}{time zone of the requested date-time interval (see \code{OlsonNames})}

\item{cols}{names or indexes of columns from the database (NULL default, returning all available columns)}

\item{quiet}{if process should be reported}

\item{rows_lim}{}

\item{original.units}{}
}
\value{
data.table with the subset of
}
\description{
Get MERRA-2 subset
}
\examples{
NA
}
\references{
<https://en.wikipedia.org/wiki/List_of_tz_database_time_zones>
}
