% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locid.R
\name{getGrid}
\alias{getGrid}
\title{Get MERRA-2 grid}
\usage{
getGrid(
  type = "points",
  class = "sp",
  locid = NULL,
  lon = c(-180, 180),
  lat = c(-90, 90)
)
}
\arguments{
\item{type}{type of grid-data to return, spatial points ("points") or polygons ("poly")}

\item{class}{class of object to return, spatial ("sp"), simple feature ("sf"), or data.frame ("df")}

\item{lon}{range of longitude}

\item{lat}{range of latitude}
}
\value{
Returns object with MERRA-2 grid information
}
\description{
Get MERRA-2 grid
}
\examples{
getGrid()
getGrid("poly", "sf")
getGrid(lon = c(-70, -60), lat = c(30, 40), class = "df")
}
