% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{merra2}
\alias{merra2}
\alias{get_merra2}
\alias{set_merra2}
\alias{check_merra2}
\title{Get, Set, or Check content of the Directory with the MERRA2 database Subset}
\usage{
get_merra2()

set_merra2(path, verbose = TRUE)

check_merra2(path = get_merra2(), detailed = FALSE)

merra2()
}
\arguments{
\item{path}{A character string of the path (default "~/data/merra2_subset")}

\item{detailed}{logical, if a list of files should be returned.}
}
\value{
\code{get_merra2} returns a character string or NULL if the directory is not set.

\code{set_merra2} returns the currently set directory before the change, invisibly.

\code{check_merra2} returns number or list of data-files in the \code{merra2} directory.
}
\description{
\code{merra2} is a subset of MERRA2 database with global

\code{get_merra2} returns the current filepath to the MERRA2 data subset;

\code{set_merra2} is used to set the directory of the database;

\code{check_merra2} checks the content of the \code{merra2} directory, and lists data-files.
}
\references{
Bosilovich, M. G., R. Lucchesi, and M. Suarez, 2016: MERRA-2: File Specification.
GMAO Office Note No. 9 (Version 1.1), 73 pp,
available from http://gmao.gsfc.nasa.gov/pubs/office_notes

Global Modeling and Assimilation Office (GMAO) (2015), MERRA-2 tavg1_2d_rad_Nx: 2d,1-Hourly,Time-Averaged,Single-Level,Assimilation,Radiation Diagnostics 0.625 x 0.5 degree V5.12.4 (M2T1NXRAD) at GES DISC. Accessed: 2019-2020 at DOI: 10.5067/Q9QMY5PBNV1T

Global Modeling and Assimilation Office (GMAO) (2015), MERRA-2 tavg1_2d_slv_Nx: 2d,1-Hourly,Time-Averaged,Single-Level,Assimilation,Single-Level Diagnostics V5.12.4 (M2T1NXSLV) at GES DISC. Accessed: 2019-2020 at DOI: 10.5067/VJAFPLI1CSIV
}
