% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merra_subset.R
\name{merra2_sample}
\alias{merra2_sample}
\title{Get a example of MERRA2 data for a particular month(s)}
\usage{
merra2_sample(month = 1:12, add.coord = FALSE)
}
\arguments{
\item{month}{integer vector with month numbers (1 to 12)}
}
\value{
An example of MERRA-2 data set for a requested month, 21st day, 2010. 
The returned data.table has the same format as returned by \code{get_merra2_subset()}.
}
\description{
Get a example of MERRA2 data for a particular month(s)
}
\examples{
 merra2_sample()
 merra2_sample(2:3)
}
