% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra2ools-package.R
\docType{package}
\name{merra2ools-package}
\alias{merra2ools}
\alias{merra2ools-package}
\title{merra2ools: R-tools to process MERRA-2 data (NASA) with focus on renewable energy}
\description{
A toolbox for evaluation of hourly availability of solar and wind energy potential globally, based on MERRA-2 satellite reanalysis data subset (requires the data download from: https://doi.org/10.5061/dryad.v41ns1rtt).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://energyRt.github.io/merra2ools/}
  \item Report bugs at \url{https://github.com/energyRt/merra2ools/issues}
}

}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@gmail.com}

Authors:
\itemize{
  \item Shuo Gao \email{gaoshuo@cet.net.cn}
}

}
\keyword{internal}
