% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_position.R
\name{plane_of_array}
\alias{plane_of_array}
\alias{fPOA}
\title{Plane-Of-Array (POA) isotropic irradiance model}
\usage{
plane_of_array(AOI, GHI, DNI, DHI, albedo, array.tilt)

fPOA(AOI, GHI, DNI, DHI, albedo, array.tilt)
}
\arguments{
\item{AOI}{Angle of Incidence, degrees}

\item{GHI}{Global Horizontal Irradiance (\mjseqn{W/m^2})}

\item{DNI}{Direct Normal Irradiance (\mjseqn{W/m^2})}

\item{DHI}{Diffuse Horizontal Irradiance (\mjseqn{W/m^2})}

\item{albedo}{ground-reflected portion of the POA irradiance (\mjseqn{W/m^2})}

\item{array.tilt}{the PV tilt angle, degrees}
}
\value{

}
\description{
Plane-Of-Array (POA) isotropic irradiance model
}
\details{
\loadmathjax
  \mjsdeqn{I_{POA} = I_{POA,b} + I_{POA,d} + I_{POA,g}}
where:
    \itemize{
      \item \mjseqn{I_{POA} \textrm{ - the plane-of-array irradiance } (W/m^2)}
      \item \mjseqn{I_{POA,b} \textrm{ - the beam irradiance that hits the array } (W/m^2)}
        \mjsdeqn{I_{POA,b} = DNI\times\cos{(AOI)}}
      \item \mjseqn{I_{POA,d} \textrm{ - the sky-diffuse portion of the POA irradiance } (W/m^2)}
        \mjsdeqn{I_{POA,d} = DHI\times\frac{1+\cos{(array.tilt)}}{2}}
      \item \mjseqn{I_{POA,g} \textrm{ - the ground-reflected portion of the POA irradiance } (W/m^2)}
        \mjsdeqn{I_{POA,g} = GHI\times{albedo}\times\frac{1-\cos{(array.tilt)}}{2}}

}
}
\examples{
NA
}
