% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{poa_irradiance}
\alias{poa_irradiance}
\alias{fPOA}
\title{Plane-Of-Array (POA) isotropic irradiance model}
\usage{
poa_irradiance(
  x,
  array.type = "fl",
  suffix = TRUE,
  AOI = "AOI",
  GHI = "SWGDN",
  DNI = "DNI",
  DHI = "DHI",
  ALBEDO = "ALBEDO",
  array.tilt = "array.tilt",
  zenith = "zenith",
  tilt.param = tilt.param.default(),
  keep.all = FALSE,
  verbose = getOption("merra2.verbose"),
  ...
)

fPOA(
  x,
  array.type = "all",
  suffix = TRUE,
  UTC = "UTC",
  yday = "yday",
  hour = "hour",
  lon = "lon",
  lat = "lat",
  integral_steps = 3,
  tilt.param = tilt.param.default(),
  keep.all = FALSE,
  verbose = getOption("merra2.verbose")
)
}
\arguments{
\item{x}{}

\item{array.type}{}

\item{suffix}{}

\item{AOI}{Angle of Incidence, degrees}

\item{GHI}{Global Horizontal Irradiance (\mjseqn{W/m^2})}

\item{DNI}{Direct Normal Irradiance (\mjseqn{W/m^2})}

\item{DHI}{Diffuse Horizontal Irradiance (\mjseqn{W/m^2})}

\item{array.tilt}{the PV tilt angle, degrees}

\item{keep.all}{}

\item{verbose}{}

\item{UTC}{}

\item{yday}{}

\item{hour}{}

\item{lon}{}

\item{lat}{}

\item{albedo}{ground-reflected portion of the POA irradiance (\mjseqn{W/m^2})}
}
\value{

}
\description{
Plane-Of-Array (POA) isotropic irradiance model
}
\details{
\loadmathjax
  \mjsdeqn{I_{POA} = I_{POA,b} + I_{POA,d} + I_{POA,g}}
where:
    \itemize{
      \item \mjseqn{I_{POA} \textrm{ - the plane-of-array irradiance } (W/m^2)}
      \item \mjseqn{I_{POA,b} \textrm{ - the beam irradiance that hits the array } (W/m^2)}
        \mjsdeqn{I_{POA,b} = DNI\times\cos{(AOI)}}
      \item \mjseqn{I_{POA,d} \textrm{ - the sky-diffuse portion of the POA irradiance } (W/m^2)}
        \mjsdeqn{I_{POA,d} = DHI\times\frac{1+\cos{(array.tilt)}}{2}}
      \item \mjseqn{I_{POA,g} \textrm{ - the ground-reflected portion of the POA irradiance } (W/m^2)}
        \mjsdeqn{I_{POA,g} = GHI\times{albedo}\times\frac{1-\cos{(array.tilt)}}{2}}

}
}
\examples{
NA
}
