% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{pv_array_position}
\alias{pv_array_position}
\title{Photovoltaic Solar Panel Orientation and Performance Models}
\usage{
pv_array_position(
  x,
  array.type = "fl",
  suffix = TRUE,
  lat = "lat",
  azimuth_Q = "azimuth_Q",
  zenith = "zenith",
  verbose = getOption("merra2.verbose"),
  tilt.param = tilt.param.default()
)
}
\arguments{
\item{x}{data.frame object with MERRA-2 subset}

\item{array.type}{type of tracking ()}

\item{suffix}{}

\item{lat}{latitude of PV location (\mjseqn{-90 \leq lat \leq 90})}

\item{azimuth_Q}{solar azimuth angle for the PV location (\mjseqn{0 \leq azimuth < 360})}

\item{zenith}{solar zenith angle for the PV location (\mjseqn{0 \leq azimuth \leq 90})}

\item{verbose}{}

\item{tilt.param}{}
}
\description{
Photovoltaic Solar Panel Orientation and Performance Models
}
\details{
\loadmathjax
\describe{
  \item{Fixed PV Panel (\mjseqn{*.fl})}
       {South-facing fixed solar PV with the tilted angle equal to the site's latitude.}
  
    \itemize{
      \item PV Tilted Angle (\mjseqn{array.tilt}, in degrees)
        \mjsdeqn{array.tilt = latitude}
      \item PV Azimuth Angle (\mjseqn{array.azimuth}, in degrees)
        \mjsdeqn{array.azimuth = 180}
    }
  
  \item{Horizontal Single-Axis PV Tracker (\mjseqn{*.th})}
       {A horizontal single-axis PV tracker with its axis in align 
       with the meridian direction and parallel to the ground.}
    \itemize{
      \item PV tilted angle under the optimal rotation strategy (\mjseqn{array.tilt}, in degrees)
        \mjsdeqn{array.tilt = \arctan{\big(\tan{(zenith)}\cos{(azimuth-array.azimuth)}\big)}}
      \item PV Azimuth Angle (\mjseqn{array.azimuth}, in degrees)
        \mjsdeqn{array.azimuth = 90 \textrm{ if } azimuth < 180}
        \mjsdeqn{array.azimuth = 270 \textrm{ if } azimuth \geq180}
    }
  
  \item{Vertical Single-Axis Tracker (\mjseqn{*.tv})}
       {A vertical single-axis PV tracker with its axis normal to the ground.}
    \itemize{
      \item PV tilted angle under the optimal rotation strategy (\mjseqn{array.tilt}, in degrees)
        \mjsdeqn{array.tilt = latitude}
      \item PV Azimuth Angle (\mjseqn{array.azimuth}, in degrees)
        \mjsdeqn{array.azimuth = azimuth}
    }
    
  \item{Tilted Single-Axis Tracker (\mjseqn{*.tl})}
       {A single-axis PV tracker with its axis parallel to the meridian direction and the axis tilted angle equal to the site’s latitude.}
    \itemize{
      \item PV tilted angle under the optimal rotation strategy (\mjseqn{array.tilt}, in degrees)
        \mjsdeqn{array.tilt = \arctan{\big(\frac{\tan{(zenith)}}
                {\cos{(array.azimuth-180)}}\big)}+\delta\pi}
        where \mjseqn{\delta=0 \textrm{ when }  90 \leq array.azimuth \leq 270, 
              \textrm{ otherwise }\delta=1}
      \item PV Azimuth Angle (\mjseqn{array.azimuth}, in degrees)
        \mjsdeqn{array.azimuth = 180(1+\sigma)+\Delta tilt}
        where
       \mjsdeqn{\sigma = \begin{cases}
           1& & {\Delta tilt < 0, azimuth \geq 180}\newline
           0& & {\Delta tilt \times (azimuth-180) \geq 0}\newline
           -1& & {\Delta tilt > 0, azimuth < 180}
           \end{cases}}
        \mjsdeqn{\Delta tilt = \arctan{\frac{\sin{(zenith)}
                 \sin{(azimuth-180)}}{\cos{(\beta)}\sin{(latitude)}}}}
        \mjsdeqn{\cos{(\beta)} = \cos{(zenith)}\cos{(latitude)}+
                 \sin{(zenith)}\sin{(latitude)}\cos{(azimuth-180)}}
        
           }
        
  \item{Dual-Axis Tracker (\mjseqn{*.td})}
       {A dual-axis PV tracker.}
  
    \itemize{
      \item PV Tilted Angle under the Optimal Rotation Strategy (\mjseqn{array.tilt}, in degrees)
        \mjsdeqn{array.tilt = zenith}
      \item PV Azimuth Angle under the Optimal Rotation Strategy (\mjseqn{array.azimuth}, in degrees)
        \mjsdeqn{array.azimuth = azimuth}
    }
    
 }
}
\examples{
NA
}
