% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{pv_array_types}
\alias{pv_array_types}
\title{List tracking system types}
\usage{
pv_array_types(array.type = "all")
}
\value{

}
\description{
List tracking system types
}
\examples{
pv_array_types()
pv_array_types("fl")
}
