% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra_subset.R
\name{read_merra_file}
\alias{read_merra_file}
\title{Reads one file from MERRA-2 subset for a given year and month}
\usage{
read_merra_file(YYYYMM, path = get_merra2_dir(), original.units = TRUE)
}
\arguments{
\item{YYYYMM}{character, year and month}

\item{path}{the database path}

\item{original.units}{should the original MERRA-2 units be returned (converted from integer vectors)}
}
\value{
A data.table with the MERRA-2 subset for the given year and month.
}
\description{
Reads one file from MERRA-2 subset for a given year and month
}
\examples{
\dontrun{
read_merra_file("201912")
}
}
