% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_geometry.R
\name{solar_geometry}
\alias{solar_geometry}
\title{Solar geometry}
\usage{
solar_geometry(datetime = NULL, yday = NULL, tz = lubridate::tz(datetime))
}
\arguments{
\item{datetime}{date and time}

\item{yday}{day of a year}

\item{tz}{timezone}
}
\value{

}
\description{
Solar geometry
}
\details{
\loadmathjax
List or data.frame with estimated following solar geometry variables:
\itemize{
  \item Solar declination (\mjseqn{\theta_d})
    \mjsdeqn{\theta_d = \frac{23.45\pi}{180}\sin{\big(2\pi\frac{284+n}{365}\big)}} 
    
  \item Equation of time (\mjseqn{E_{qt}})
     \mjsdeqn{
       E_{qt} = \begin{cases}
         -14.2\sin{\big(\frac{\pi(n+7)}{111}\big)}& & {1 \leq n \leq 106}\newline 
         4.0\sin{\big(\frac{\pi(n-106)}{59}\big)}& & {107 \leq n \leq 166}\newline 
         -6.5\sin{\big(\frac{\pi(n-166)}{80}\big)}& & {167 \leq n \leq 246}\newline 
         16.4\sin{\big(\frac{\pi(n-247)}{113}\big)}& & {247 \leq n \leq 365}
         \end{cases}
       }
       
  \item Apparent solar time (\mjseqn{T_{solar}})
     \mjsdeqn{T_{solar} = T_{UTC}+\frac{E_{qt}}{60}+longitude/15}
     
  \item Hour angle (\mjseqn{\theta_{hr}})
     \mjsdeqn{\theta_{hr} = \frac{T_{solar}-12}{12}\pi}
       
  \item Zenith angle
     \mjsdeqn{zenith = \arccos\big({\sin{(latitude)}\sin{(\theta_d)}+\cos{(latitude)}\cos{(\theta_d)}\cos{(\theta_{hr})}\big)}}
     
  \item Azimuth angle
     \mjsdeqn{
       azimuth = \begin{cases}
         \arcsin{(A)}& & {A \geq 0, B \geq 0}\newline
         180-\arcsin{(A)}& & {B < 0}\newline
         360+\arcsin{(A)}& & {A < 0, B \geq 0}
         \end{cases}
     }
     where \cr
            \mjseqn{n - \text{day of the year}} \cr
            \mjseqn{A = \sin{(azimuth)} = 
                       -\frac{\sin{(\theta_{hr})}\cos{(\theta_d)}}{\sin{(zenith)}}} \cr
            \mjseqn{B = \cos{(azimuth)} = \frac{\sin{(\theta_d)}-
                        \sin{(latitude)}\cos{(zenith)}}{\cos{(latitude)}\sin{(zenith)}}}
}
}
\examples{
NA
}
