% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{tilt.param.default}
\alias{tilt.param.default}
\title{Default tilt-parameters of tracking systems}
\usage{
tilt.param.default(x = NULL)
}
\arguments{
\item{x}{}
}
\description{
Default tilt-parameters of tracking systems
}
\examples{
str(tilt.param.default())
}
