% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{wind_GIF}
\alias{wind_GIF}
\title{Quick GIF plots}
\usage{
wind_GIF(
  merra,
  name = "W50M",
  mult = 1,
  filename = "merra_wind.gif",
  nmax = 48,
  fps = 24,
  gif.width = 1920/3,
  gif.height = 1080/3,
  limits = c(0, 35),
  palette = "RdYlBu",
  direction = -1,
  legend.position = c(0.91, 0.05),
  legend.name = "m/s",
  datetime.position = c(149, 87),
  datetime.format = "\%Y-\%b-\%d, \%Hh \%Z",
  verbose = TRUE
)
}
\arguments{
\item{merra}{}

\item{name}{}

\item{mult}{}

\item{filename}{}

\item{nmax}{}

\item{fps}{}

\item{gif.width}{}

\item{gif.height}{}

\item{limits}{}

\item{legend.position}{}

\item{datetime.position}{}

\item{datetime.format}{}

\item{verbose}{}
}
\value{

}
\description{
Quick GIF plots
}
