% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{wind_plot}
\alias{wind_plot}
\title{Quick plots of wind speed and its dynamics}
\usage{
wind_plot(
  merra,
  name = "W50M",
  mult = 1,
  limits = c(0, 35),
  palette = "RdYlBu",
  direction = -1,
  legend.position = c(0.91, 0.05),
  legend.name = "m/s",
  datetime.position = c(150, 87),
  datetime.format = "\%Y-\%b-\%d, \%Hh \%Z"
)
}
\arguments{
\item{merra}{}

\item{name}{}

\item{mult}{}

\item{limits}{}

\item{legend.position}{}

\item{datetime.position}{}

\item{datetime.format}{}
}
\value{

}
\description{
Quick plots of wind speed and its dynamics
}
