% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrel_pvwatts_api.R
\name{yday2date}
\alias{yday2date}
\title{Convert the day of the year to}
\usage{
yday2date(yday, year = 2010, tz = "UTC")
}
\arguments{
\item{yday}{integer vector with days of a year (1-366)}

\item{year}{integer vector with years}

\item{tz}{character time zone name}
}
\description{
Convert the day of the year to
}
\examples{
NA
}
