% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra_subset.R
\name{add_coord}
\alias{add_coord}
\title{Add longitude and latitude of 'locid'}
\usage{
add_coord(x, replace = FALSE)
}
\arguments{
\item{x}{data.frame with `locid` column.}

\item{replace}{logical, if TRUE, existing `lon` and `lat` columns in `x` will be replaced.}
}
\value{
`x` with added/replaced `lon` and `lat` for each `locid`
}
\description{
Add longitude and latitude of 'locid'
}
\examples{
NA
}
