% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_pv_models.R
\name{angle_of_incidence}
\alias{angle_of_incidence}
\alias{fAOI}
\title{Angle of Incidence (AOI)}
\usage{
angle_of_incidence(
  x,
  array.type = "fh",
  suffix = TRUE,
  azimuth_Q = "azimuth_Q",
  zenith = "zenith",
  na.val = NA,
  zenith.max = 90,
  AOI.max = 90,
  verbose = getOption("merra2.verbose"),
  ...
)

fAOI(
  x,
  array.type = "fh",
  suffix = TRUE,
  azimuth_Q = "azimuth_Q",
  zenith = "zenith",
  na.val = NA,
  zenith.max = 90,
  AOI.max = 90,
  verbose = getOption("merra2.verbose"),
  ...
)
}
\arguments{
\item{x}{}

\item{array.type}{}

\item{suffix}{}

\item{azimuth_Q}{the solar zenith angle, degrees}

\item{zenith}{the solar azimuth angle, degrees}

\item{na.val}{}

\item{zenith.max}{}

\item{verbose}{}

\item{...}{}
}
\description{
Angle of Incidence (AOI)
}
\details{
\loadmathjax
    \mjsdeqn{AOI = \arccos\big({\cos{(zenith)}\cos{(array.tilt)}+
             \sin{(zenith)}\sin{(array.tilt)}\cos{(azimuth-array.azimuth)}\big)}}
    Though the equation returns AOI values with any given set of zenith, azimuth, 
    array.tilt, and array.azimuth, the AOI is meaningful only if 
    \mjseqn{0 \leq zenith \leq 90}, and \mjseqn{0 \leq AOI \leq 90}. 
    The former condition assures the sun is above the horizon and 
    the latter condition assures the sunlight beam is able to hit the panel
    (Source: Stackhouse et al., 2018).
}
\examples{
NA
}
