% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merra_subset.R
\name{fDate}
\alias{fDate}
\title{Title}
\usage{
fDate(year, month, day, hour, tz = "UTC")
}
\arguments{
\item{year}{}

\item{month}{}

\item{day}{}

\item{hour}{}

\item{tz}{}
}
\value{

}
\description{
Title
}
