% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locid.R
\name{get_locid}
\alias{get_locid}
\title{Get MERRA-2 grid IDs that overlay or intersect with given spatial object}
\usage{
get_locid(sp, method = "intersect", return_sp = FALSE, projString = NULL)
}
\arguments{
\item{sp}{a map-object of class SpatialPolygons or SpatialPolygonsDataFrame}

\item{method}{character, "points" or "intersect", matching spatial points with polygons or intersection of polygons respectively}

\item{return_sp}{logical, if TRUE, Spatial object will be returned}

\item{projString}{projection string for return object (in the case of return_sp = TRUE)}
}
\value{
when return_sp = FALSE, an integer vector with location IDs (locid) is returned
}
\description{
Get MERRA-2 grid IDs that overlay or intersect with given spatial object
}
\examples{
iceland_sp <- rnaturalearth::ne_states(country = "iceland")
lid2_sp <- get_locid(iceland_sp, method = "intersect", return_sp = TRUE)
sp::plot(lid2_sp, col = "wheat", axes = TRUE, main = "Iceland")
lid1 <- get_locid(iceland_sp, method = "points")
lid2 <- get_locid(iceland_sp, method = "intersect", )
lid1_df <- getGrid(locid = lid1, class = "df")
lid2_df <- getGrid(locid = lid2, class = "df")
points(lid2_df$lon, lid2_df$lat, col = "red")
points(lid1_df$lon, lid1_df$lat, col = "red", pch = 16)

}
