% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locid.R
\name{get_locid}
\alias{get_locid}
\title{Get MERRA-2 grid IDs which overlaps with the given spatial object}
\usage{
get_locid(
  x,
  method = "polygons",
  add_poles_points = F,
  return_grid = FALSE,
  ...
)
}
\arguments{
\item{x}{spatial (`sf`) object.}

\item{method}{}

\item{...}{ignored}
}
\description{
Get MERRA-2 grid IDs which overlaps with the given spatial object
}
