% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{gif_merra}
\alias{gif_merra}
\title{Quick GIF plots}
\usage{
gif_merra(
  x,
  FUN = "plot_merra",
  timestamp.variable = "UTC",
  nmax = 24 * 12,
  fps = 12,
  gif.width = 576,
  gif.height = 360,
  filename = NULL,
  filename.prefix = "merra_",
  dirname = ".",
  verbose = getOption("merra2.verbose"),
  ...
)
}
\arguments{
\item{x}{}

\item{FUN}{}

\item{nmax}{}

\item{fps}{}

\item{gif.width}{}

\item{gif.height}{}

\item{filename}{}

\item{dirname}{}

\item{verbose}{}

\item{...}{}
}
\description{
Quick GIF plots
}
\examples{
NA
}
