% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{plot_merra}
\alias{plot_merra}
\title{Quick plots of wind speed and its dynamics}
\usage{
plot_merra(
  x,
  name = "W50M",
  scale = 1,
  intercept = 0,
  limits = c(0, 35),
  palette = "RdYlBu",
  direction = -1,
  na.value = "#040434",
  legend.position = c(0.91, 0.05),
  legend.name = "",
  timestamp.variable = "UTC",
  timestamp.format = "\%Y-\%b-\%d, \%Hh \%Z",
  timestamp.position = c(150, 87),
  expand.x = c(0, 0),
  expand.y = c(0, 0),
  map.border = "lightgrey",
  map.border.size = 0.25,
  ...
)
}
\arguments{
\item{x}{}

\item{name}{}

\item{scale}{}

\item{intercept}{}

\item{limits}{}

\item{palette}{}

\item{direction}{}

\item{legend.position}{}

\item{legend.name}{}

\item{timestamp.format}{}

\item{timestamp.position}{}

\item{expand.x}{}

\item{expand.y}{}

\item{...}{}
}
\description{
The function is in progress, will be included in the package release.
}
\examples{
NA
}
